pyproject.toml
[project]
name = "crewaidocsmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.13",
    "beautifulsoup4>=4.13.4",
    "lru-dict>=1.3.0",
    "mcp[cli]>=1.10.1",
    "python-dotenv>=1.1.1",
    "whoosh>=2.7.4",
]
